fix log print crash on 32-bits systems

+++ src/transport/xqc_engine.c
@@ -322,7 +322,7 @@ xqc_engine_schedule_reset(xqc_engine_t *engine,
 
     uint32_t hash = xqc_murmur_hash2((unsigned char*)peer_addr, peer_addrlen);
     hash = hash % XQC_RESET_CNT_ARRAY_LEN;
-    xqc_log(engine->log, XQC_LOG_DEBUG, "|hash:%ud|cnt:%ud|", hash, engine->reset_sent_cnt[hash]);
+    xqc_log(engine->log, XQC_LOG_DEBUG, "|hash:%ud|cnt:%ud|", hash, (unsigned int)engine->reset_sent_cnt[hash]);
 
     if (engine->reset_sent_cnt[hash] < 2) {
         engine->reset_sent_cnt[hash]++;
